refactor: rename and move PromoBanner to shared components
