+++ src/components/SharedComponents/PromoBanner/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./PromoBanner";
